No issue. Fix some bugs regading the loading of the catalog - Add catalog declaration to all test scripts. - Update reference data to DKPro Core 1.9.0-SNAPSHOT changes

+++ openminted-script-groovy/src/main/groovy/eu/openminted/script/groovy/internal/gate/GateComponentInstance.groovy
@@ -62,27 +62,27 @@ class GateComponentInstance implements ComponentInstance {
 		if (!(document.data instanceof gate.Document)) {
 			throw new IllegalArgumentException("Cannot process $document");
 		}
-		if(delegate instanceof LanguageAnalyser)
+        
+		if (delegate instanceof LanguageAnalyser)
 		{
 			delegate.setDocument(document.data);
 			delegate.execute();
-		}else if(delegate instanceof DocumentExporter){
+		}
+        else if(delegate instanceof DocumentExporter) {
 			def out;
-			if(parameters)
-			{
-				if(parameters['targetLocation'])
-				{
-					out = new File(parameters['targetLocation']+"/output.txt");
-					if(parameters["overwrite"])
-					{
+			if (parameters) {
+				if(parameters['targetLocation']) {
+					out = new File(parameters['targetLocation']);
+					if(parameters["overwrite"] && out.exists()) {
 						out.delete();						
 					}					
-					out.createNewFile();
+                    out.getParentFile().mkdirs();
 				}
 			}
 
-			if(!out)
+			if (!out) {
 				out = System.out;
+			}
 
 			delegate.export(document.data,out);
 		}
